Rename example to ./src/examples and add basic example
